# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.CreateEventRequest:
    properties:
      duration_minutes:
        minimum: 1
        type: integer
      end_time:
        type: string
      name:
        type: string
      passing_grade:
        maximum: 100
        minimum: 0
        type: number
      start_time:
        type: string
    required:
    - duration_minutes
    - end_time
    - name
    - passing_grade
    - start_time
    type: object
...
  domain.Event:
    properties:
      created_at:
        type: string
      duration_minutes:
        type: integer
      end_time:
        type: string
      id:
        type: string
      name:
        type: string
      passing_grade:
        type: number
      start_time:
        type: string
    type: object
...
  response.ErrorDetail:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      errors:
        items:
          $ref: '#/definitions/response.ErrorDetail'
        type: array
      message:
        example: Terjadi kesalahan
        type: string
      success:
        example: false
        type: boolean
    type: object
  response.Meta:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      total_pages:
        type: integer
      total_records:
        type: integer
    type: object
  response.PaginatedResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        example: true
        type: boolean
    type: object
  response.SuccessResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: Operasi berhasil
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    name: Dev Team Pramuka CAT
    url: https://github.com/odealidj/pramuka-CAT
  description: REST API untuk platform Computer Assisted Test (CAT) Pramuka.
  termsOfService: http://swagger.io/terms/
  title: Pramuka CAT API
  version: "1.0"
paths:
...
  /admin/events:
    get:
      parameters:
      - description: Halaman
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/response.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.Event'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Daftar Event Ujian
      tags:
      - Admin - Event
    post:
      consumes:
      - application/json
      parameters:
      - description: Data Event
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.Event'
              type: object
      security:
      - BearerAuth: []
      summary: Buat Event Ujian
      tags:
      - Admin - Event
  /admin/events/{id}/export:
    get:
      description: Download rekap nilai peserta dalam format Excel (.xlsx) atau PDF
      parameters:
      - description: UUID Event
        in: path
        name: id
        required: true
        type: string
      - description: 'Format file: excel (default) atau pdf'
        in: query
        name: format
        type: string
      produces:
      - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      responses:
        "200":
          description: OK
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            type: file
        "400":
          description: Bad Request
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Export Laporan Peserta
      tags:
      - Admin - Event
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token JWT dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"